refactor(payment): CHECKOUT-3791 Move payment strategy options interface
