+++ src/payment/strategies/braintree/braintree-visacheckout-payment-initialize-options.ts
@@ -1,0 +1,22 @@
+/**
+ * A set of options that are required to initialize the Visa Checkout payment
+ * method provided by Braintree.
+ *
+ * If the customer chooses to pay with Visa Checkout, they will be asked to
+ * enter their payment details via a modal. You can hook into events emitted by
+ * the modal by providing the callbacks listed below.
+ */
+export default interface BraintreeVisaCheckoutPaymentInitializeOptions {
+    /**
+     * A callback that gets called when Visa Checkout fails to initialize or
+     * selects a payment option.
+     *
+     * @param error - The error object describing the failure.
+     */
+    onError?(error: Error): void;
+
+    /**
+     * A callback that gets called when the customer selects a payment option.
+     */
+    onPaymentSelect?(): void;
+}
