+++ src/payment/strategies/braintree/braintree-visacheckout-payment-strategy.ts
@@ -141,26 +141,3 @@ export default class BraintreeVisaCheckoutPaymentStrategy implements PaymentStra
         throw error;
     }
 }
-
-/**
- * A set of options that are required to initialize the Visa Checkout payment
- * method provided by Braintree.
- *
- * If the customer chooses to pay with Visa Checkout, they will be asked to
- * enter their payment details via a modal. You can hook into events emitted by
- * the modal by providing the callbacks listed below.
- */
-export interface BraintreeVisaCheckoutPaymentInitializeOptions {
-    /**
-     * A callback that gets called when Visa Checkout fails to initialize or
-     * selects a payment option.
-     *
-     * @param error - The error object describing the failure.
-     */
-    onError?(error: Error): void;
-
-    /**
-     * A callback that gets called when the customer selects a payment option.
-     */
-    onPaymentSelect?(): void;
-}
