+++ src/payment/strategies/braintree/index.ts
@@ -11,4 +11,5 @@ export { default as BraintreeSDKCreator } from './braintree-sdk-creator';
 export { default as createBraintreePaymentProcessor } from './create-braintree-payment-processor';
 export { default as createBraintreeVisaCheckoutPaymentProcessor } from './create-braintree-visacheckout-payment-processor';
 export { default as VisaCheckoutScriptLoader } from './visacheckout-script-loader';
-export { default as BraintreeVisaCheckoutPaymentStrategy, BraintreeVisaCheckoutPaymentInitializeOptions } from './braintree-visacheckout-payment-strategy';
+export { default as BraintreeVisaCheckoutPaymentStrategy } from './braintree-visacheckout-payment-strategy';
+export { default as BraintreeVisaCheckoutPaymentInitializeOptions } from './braintree-visacheckout-payment-initialize-options';
