+++ src/payment/strategies/credit-card/index.ts
@@ -1,0 +1,1 @@
+export { default as CreditCardPaymentStrategy } from './credit-card-payment-strategy';
