+++ src/payment/strategies/index.ts
@@ -1,28 +1,1 @@
-export { default as CreditCardPaymentStrategy } from './credit-card-payment-strategy';
-export { default as LegacyPaymentStrategy } from './legacy-payment-strategy';
-export { default as NoPaymentDataRequiredPaymentStrategy } from './no-payment-data-required-strategy';
-export { default as OfflinePaymentStrategy } from './offline-payment-strategy';
-export { default as OffsitePaymentStrategy } from './offsite-payment-strategy';
 export { default as PaymentStrategy } from './payment-strategy';
-export { default as SagePayPaymentStrategy } from './sage-pay-payment-strategy';
-
-export { AfterpayPaymentStrategy } from './afterpay';
-export { AmazonPayPaymentStrategy, AmazonPayPaymentInitializeOptions } from './amazon-pay';
-export {
-    BraintreeCreditCardPaymentStrategy,
-    BraintreePaymentInitializeOptions,
-    BraintreePaypalPaymentStrategy,
-    BraintreeVisaCheckoutPaymentStrategy,
-    BraintreeVisaCheckoutPaymentInitializeOptions
-} from './braintree';
-export {
-    GooglePayPaymentStrategy,
-    GooglePayPaymentInitializeOptions,
-    GooglePayPaymentProcessor
-} from './googlepay';
-export { KlarnaPaymentStrategy, KlarnaPaymentInitializeOptions } from './klarna';
-export { PaypalExpressPaymentStrategy, PaypalExpressPaymentInitializeOptions, PaypalProPaymentStrategy } from './paypal';
-export { ChasePayPaymentStrategy, ChasePayInitializeOptions } from './chasepay';
-export { SquarePaymentStrategy, SquarePaymentInitializeOptions } from './square';
-export { MasterpassPaymentStrategy, MasterpassPaymentInitializeOptions } from './masterpass';
-export { WepayPaymentStrategy } from './wepay';
