+++ src/payment/strategies/klarna/klarna-payment-initialize-options.ts
@@ -1,0 +1,23 @@
+import { KlarnaLoadResponse } from './klarna-credit';
+
+/**
+ * A set of options that are required to initialize the Klarna payment method.
+ *
+ * When Klarna is initialized, a widget will be inserted into the DOM. The
+ * widget has a list of payment options for the customer to choose from.
+ */
+export default interface KlarnaPaymentInitializeOptions {
+    /**
+     * The ID of a container which the payment widget should insert into.
+     */
+    container: string;
+
+    /**
+     * A callback that gets called when the widget is loaded and ready to be
+     * interacted with.
+     *
+     * @param response - The result of the initialization. It indicates whether
+     * or not the widget is loaded successfully.
+     */
+    onLoad?(response: KlarnaLoadResponse): void;
+}
