+++ src/payment/strategies/legacy/index.ts
@@ -1,0 +1,1 @@
+export { default as LegacyPaymentStrategy } from './legacy-payment-strategy';
