+++ src/payment/strategies/no-payment/index.ts
@@ -1,0 +1,1 @@
+export { default as NoPaymentDataRequiredPaymentStrategy } from './no-payment-data-required-strategy';
