+++ src/payment/strategies/no-payment/no-payment-data-required-strategy.ts
@@ -1,11 +1,11 @@
 import { omit } from 'lodash';
 
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import { OrderActionCreator, OrderRequestBody } from '../../order';
-import { OrderFinalizationNotRequiredError } from '../../order/errors';
-import { PaymentInitializeOptions, PaymentRequestOptions } from '../payment-request-options';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import { OrderActionCreator, OrderRequestBody } from '../../../order';
+import { OrderFinalizationNotRequiredError } from '../../../order/errors';
+import { PaymentInitializeOptions, PaymentRequestOptions } from '../../payment-request-options';
 
-import PaymentStrategy from './payment-strategy';
+import PaymentStrategy from '../payment-strategy';
 
 export default class NoPaymentDataRequiredPaymentStrategy implements PaymentStrategy {
     constructor(
