+++ src/payment/strategies/offline/index.ts
@@ -1,0 +1,1 @@
+export { default as OfflinePaymentStrategy } from './offline-payment-strategy';
