+++ src/payment/strategies/offline/offline-payment-strategy.ts
@@ -1,9 +1,9 @@
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import { OrderActionCreator, OrderRequestBody } from '../../order';
-import { OrderFinalizationNotRequiredError } from '../../order/errors';
-import { PaymentInitializeOptions, PaymentRequestOptions } from '../payment-request-options';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import { OrderActionCreator, OrderRequestBody } from '../../../order';
+import { OrderFinalizationNotRequiredError } from '../../../order/errors';
+import { PaymentInitializeOptions, PaymentRequestOptions } from '../../payment-request-options';
 
-import PaymentStrategy from './payment-strategy';
+import PaymentStrategy from '../payment-strategy';
 
 export default class OfflinePaymentStrategy implements PaymentStrategy {
     constructor(
