+++ src/payment/strategies/offsite/index.ts
@@ -1,0 +1,1 @@
+export { default as OffsitePaymentStrategy } from './offsite-payment-strategy';
