+++ src/payment/strategies/sage-pay/index.ts
@@ -1,0 +1,1 @@
+export { default as SagePayPaymentStrategy } from './sage-pay-payment-strategy';
