+++ src/payment/strategies/sage-pay/sage-pay-payment-strategy.ts
@@ -1,15 +1,15 @@
 import { some } from 'lodash';
 
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import { RequestError } from '../../common/error/errors';
-import { OrderActionCreator, OrderRequestBody } from '../../order';
-import { OrderFinalizationNotRequiredError } from '../../order/errors';
-import { PaymentArgumentInvalidError } from '../errors';
-import PaymentActionCreator from '../payment-action-creator';
-import { PaymentInitializeOptions, PaymentRequestOptions } from '../payment-request-options';
-import * as paymentStatusTypes from '../payment-status-types';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import { RequestError } from '../../../common/error/errors';
+import { OrderActionCreator, OrderRequestBody } from '../../../order';
+import { OrderFinalizationNotRequiredError } from '../../../order/errors';
+import { PaymentArgumentInvalidError } from '../../errors';
+import PaymentActionCreator from '../../payment-action-creator';
+import { PaymentInitializeOptions, PaymentRequestOptions } from '../../payment-request-options';
+import * as paymentStatusTypes from '../../payment-status-types';
 
-import PaymentStrategy from './payment-strategy';
+import PaymentStrategy from '../payment-strategy';
 
 export default class SagePayPaymentStrategy implements PaymentStrategy {
     constructor(
